+++ sdk/src/main/scala/com/ubirch/defaults/TokenApi.scala
@@ -2,30 +2,12 @@ package com.ubirch.defaults
 
 import java.util.UUID
 
-import com.google.inject.AbstractModule
-import com.typesafe.config.Config
 import com.ubirch.api._
-import com.ubirch.utils.{ ConfigProvider, InjectorHelper }
 import monix.eval.Task
-import org.json4s.Formats
 
 import scala.util.{ Failure, Try }
 
-object TokenApi extends TokenManager {
-
-  final private val injector: InjectorHelper = new InjectorHelper(List(new AbstractModule {
-    override def configure(): Unit = {
-      bind(classOf[JsonConverterService]).to(classOf[DefaultJsonConverterService])
-      bind(classOf[TokenPublicKey]).to(classOf[DefaultTokenPublicKey])
-      bind(classOf[TokenVerification]).to(classOf[DefaultTokenVerification])
-      bind(classOf[ExternalStateGetter]).to(classOf[DefaultExternalGetter])
-      bind(classOf[ExternalStateVerifier]).to(classOf[DefaultStateVerifier])
-
-      bind(classOf[Config]).toProvider(classOf[ConfigProvider])
-      bind(classOf[Formats]).toProvider(classOf[JsonFormatsProvider])
-      ()
-    }
-  })) {}
+object TokenApi extends TokenManager with Binding {
 
   final private val tokenVerification: TokenVerification = injector.get[TokenVerification]
   final private val externalStateVerifier: ExternalStateVerifier = injector.get[ExternalStateVerifier]
